Authors – should be able to register with basic info which includes email address. There should be open registration for authors where they should complete email validation to complete it
Authors can add one or many articles

Login, and forgot password pages should be implemented so that registered users can log in and add articles.
The application should have an admin user to monitor Users and Articles, they should be able to delete Authors and articles. No need for open registration.

Articles – belong to the author